Make big component into little pieces.
The SharedPanelAccountItem Component was getting quite
complicated and was starting to get difficult to reason about.
This commit splits the component into discrete and reusable
parts.
